calendar logic added
