+++ lib/widgets/home_widgets/DashboardsWidgets.dart
@@ -1,16 +1,20 @@
 // ignore_for_file: unnecessary_this, curly_braces_in_flow_control_structures
 
 import 'package:flutter/material.dart';
+import 'package:personal_nutrition_control/models/DashboardModels.dart';
 import 'package:personal_nutrition_control/models/DayModel.dart';
-import 'package:personal_nutrition_control/models/UserModel.dart';
-import 'package:personal_nutrition_control/providers/DayProvider.dart';
 import 'package:personal_nutrition_control/providers/UserProvider.dart';
 import 'package:provider/provider.dart';
 
 import 'package:syncfusion_flutter_charts/charts.dart';
 
 class DiaryIndicators extends StatefulWidget {
-  const DiaryIndicators({super.key});
+  final DayModel dayToView;
+
+  const DiaryIndicators({
+    required this.dayToView,
+    super.key
+  });
 
   @override
   State<DiaryIndicators> createState() => _DiaryIndicatorsState();
@@ -30,36 +34,27 @@ class _DiaryIndicatorsState extends State<DiaryIndicators> {
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<DayProvider>(
-      builder: (context, dayProvider, child){
-        UserModel? user = Provider.of<UserProvider>(context, listen: false).user;
-        DayModel? actualDay = dayProvider.actualDay;
-        if(user == null || actualDay == null)
-          return Container();
 
-        DateTime now = DateTime.now();
-        final DayModel? day = dayProvider.days.firstWhere((d) => d!.date.contains('${now.year}-${now.month}-${now.day}'));
+    int targetCalories = Provider.of<UserProvider>(context, listen: false).user?.targetCalories ?? 2000;
+    double size = MediaQuery.of(context).size.height;
 
-        if (day == null)
-          return Text('Día no encontrado');
-        // Actualiza tu interfaz de usuario con los datos del día específico
-        return Row(
-          children: [
-            CaloriesIndicator(
-              tooltipBehavior: _proteinsTooltipBehavior,
-              actualValue: day.caloriesConsumed.toInt(),
-              maximumValue: user.targetCalories,
-              size: 200,
-            ),
-            MacronutrientsIndicator(
-              tooltipBehavior: _macrosTooltipBehavior,
-              totalProteins: day.proteinsConsumed,
-              totalCarbs: day.carbsConsumed,
-              totalFats: day.fatsConsumed,
-            )
-          ],
-        );
-      },
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceAround,
+      children: [
+        CaloriesIndicator(
+          tooltipBehavior: _proteinsTooltipBehavior,
+          actualValue: widget.dayToView.caloriesConsumed.toInt(),
+          maximumValue: targetCalories,
+          size: size * 0.2,
+        ),
+        MacronutrientsIndicator(
+          size: size * 0.2,
+          tooltipBehavior: _macrosTooltipBehavior,
+          totalProteins: widget.dayToView.proteinsConsumed,
+          totalCarbs: widget.dayToView.carbsConsumed,
+          totalFats: widget.dayToView.fatsConsumed,
+        )
+      ],
     );
   }
 }
@@ -86,9 +81,9 @@ class CaloriesIndicator extends StatelessWidget {
       height: this.size,
       width: this.size,
       child: SfCircularChart(
+        margin: EdgeInsets.all(0),
         series: <CircularSeries>[
           RadialBarSeries<ArcProgressIndicatorData, String>(
-            dataSource: [ ArcProgressIndicatorData("Calories", this.actualValue ) ],
             xValueMapper: (data, _) => data.x,
             yValueMapper: (data, _) => data.y,
             pointColorMapper: (data, _) => Colors.deepOrangeAccent,
@@ -96,12 +91,15 @@ class CaloriesIndicator extends StatelessWidget {
             trackOpacity: 0.1,
             cornerStyle: CornerStyle.bothCurve,
             maximumValue: this.maximumValue.toDouble(),
+            dataSource: [
+              ArcProgressIndicatorData("Calories", this.actualValue )
+            ],
           ),
         ],
         tooltipBehavior: this.tooltipBehavior,
         annotations: [
           CircularChartAnnotation(
-              widget: Text('Proteins')
+            widget: const Text('Proteins')
           )
         ],
       ),
@@ -114,12 +112,14 @@ class MacronutrientsIndicator extends StatelessWidget {
   final double totalProteins;
   final double totalFats;
   final double totalCarbs;
+  final double size;
   final TooltipBehavior tooltipBehavior;
 
   const MacronutrientsIndicator({
     required this.totalProteins,
     required this.totalFats,
     required this.totalCarbs,
+    required this.size,
     required this.tooltipBehavior,
     super.key
   });
@@ -139,41 +139,26 @@ class MacronutrientsIndicator extends StatelessWidget {
     ];
 
     return SizedBox(
-        height: 200,
-        width: 200,
-        child: SfCircularChart(
-          tooltipBehavior: this.tooltipBehavior,
-          series: <PieSeries<MacronutrientsIndicatorData, String>>[
-            PieSeries<MacronutrientsIndicatorData, String>(
-              pointColorMapper:(data,  _) => data.color,
-              xValueMapper: (data, _) => data.x,
-              yValueMapper: (data, _) => data.y,
-              dataLabelMapper: (data, _) => data.x,
-              dataSource: chartData,
-              dataLabelSettings: const DataLabelSettings(
-                connectorLineSettings: ConnectorLineSettings(
-                  length: '15',
-                  type: ConnectorType.curve
-                ),
-                isVisible: true,
-              ),
+      height: this.size,
+      width: this.size,
+      child: SfCircularChart(
+
+        margin: EdgeInsets.all(0),
+        tooltipBehavior: this.tooltipBehavior,
+        series: <PieSeries<MacronutrientsIndicatorData, String>>[
+          PieSeries<MacronutrientsIndicatorData, String>(
+            pointColorMapper:(data,  _) => data.color,
+            xValueMapper: (data, _) => data.x,
+            yValueMapper: (data, _) => data.y,
+            dataLabelMapper: (data, _) => data.x,
+            dataSource: chartData,
+            dataLabelSettings: const DataLabelSettings(
+              connectorLineSettings: ConnectorLineSettings(length: '15', type: ConnectorType.curve),
+              isVisible: true,
             ),
-          ],
-        )
+          ),
+        ],
+      )
     );
   }
-}
-
-class ArcProgressIndicatorData {
-  ArcProgressIndicatorData(this.x, this.y);
-
-  final String? x;
-  final int? y;
-}
-
-class MacronutrientsIndicatorData {
-  MacronutrientsIndicatorData(this.x, this.y, [this.color]);
-  final String x;
-  final double y;
-  final Color? color;
 }
