+++ lib/widgets/home_widgets/Searcher.dart
@@ -1,0 +1,32 @@
+
+
+import 'package:flutter/material.dart';
+
+class Searcher extends StatelessWidget {
+  const Searcher({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return InkWell(
+      onTap: (){},
+      child: Container(
+        height: 35,
+        decoration: BoxDecoration(
+          border: Border.all(color: Colors.orange, width: 1.4),
+          borderRadius: BorderRadius.circular(25),
+        ),
+        child: const Row(
+          children: [
+            Padding(
+              padding: EdgeInsets.symmetric(horizontal: 10),
+              child: Icon(Icons.search, color: Colors.grey),
+            ),
+            Text('Find Food to add', style: TextStyle(color: Colors.grey)),
+          ],
+        ),
+      ),
+    );
+  }
+}
