+++ pubspec.yaml
@@ -47,6 +47,7 @@ dependencies:
   syncfusion_flutter_charts: ^23.2.6
   syncfusion_flutter_core: ^23.2.6
   rive: ^0.12.3
+  intl: ^0.18.1
 
 dev_dependencies:
   flutter_test:
